darray: overhaul to not use memcpy; store pointers only
The previous version of DArray used memcpy to facilitate
storing stack-allocated variables into the DArray. However,
this makes it clumsy to store and access heap-allocated
items, to the point of requiring a separate version of
DArray that specifically handles only pointers to data.

While it is nice to have the ability to store copies of
stack-allocated data into a dynamic array, the more
likely use-case is to store pointers to dynamically
allocated data. To that end, DArray has been refactored
to remove the use of memcpy and replace it with the
appropriate functionality for storing generic pointers.

Also, several bugs were discovered and fixed, notably,
in darray_maybe_resize where the grow/shrink determination
was being made incorrectly, and in darray_prepend and
darray_insert where the corner case of darray->length == 0
was not being handled properly.