darray: remove darray_slice
darray_slice is being remove until the semantics of
slices can be better defined. Right now, creating a
slice from a DArray raises an important question as
to what should happen to the slice when it is no
longer of use - should the user call darray_free_all
or darray_free? If darray_free_all, then how to free
the original DArray from which the slice was created?
If darray_free, then do we just assume the original
DArray has darray_free_all invoked on it later?

Slices need to be handled more carefully. It might be
worth looking at how Go handles them.