slist,dlist: fix usage of sentinel node
Implementation and use of a sentinel node in SList and
DList was incorrect due to a poor description of the
technique in a textbook. A better description is
available in "Introduction to Algorithms", CLR.

Using a sentinel node in SList really doesn't buy all
that much in terms of simplifying the code. It maybe
makes the code slightly cleaner for insert/remove, but
not by much. It's difficult to justify the extra space
used by the sentinel, so it might disappear in a future
commit.

For DList, however, the code simplification is quite
significant and worth the extra memory overhead.