rbtree: initial commit of red-black tree data structure
The implementation herein is based on the description
given in "Introduction to Algorithms", by Cormen, et al.
(MIT Press, 1991). Cormen, et al.'s description has been
modified so that a sentinel node is not used, and the
delete algorithm relinks the successor node into the
place of the node to be deleted rather than copying the
successor node's contents. This is necessary because
external code may maintain pointers to nodes within the
tree by way of the RBTreeIterator type.