deque,pqueue,queue,stack: Remove unnecessary containers
The dynamically allocated container structs for the queue,
stack, queue, and pqueue types are not strictly necessary.
Each of these types essentially "inherits" from some base
type, encapsulating an instance of that type in a struct.
Since a pointer to a structure object, suitably converted,
points to its initial member, the extra malloc can be
eliminated at the cost of some extra casts in the code.