dlist: dlist_mergesort: speed up by ~50%
Instead of iterating from the head of the list at the
start of each inner loop iteration, added a 'next' pointer
to point to the first element of the first sublist which the
next inner loop iteration will examine. This next pointer is
incremented and set before the merge operation happens, and
thus remains unaffected by any element rearrangements effected
by the merge routine.

Still need to do a more careful analysis of the time complexity
to verify O(n * log(n)) performance.